requirements.txt
selene==2.0.0.b16
pytest==7.2.2
allure-pytest==2.13.1
requests==2.28.2
python-dotenv==1.0.0
faker==18.3.1